requirements.txt
paramiko
PyYAML
scp